\name{kequate-package}
\alias{kequate-package}
\docType{package}
\title{
The Kernel Method of Test Equating
}
\description{
The kernel equating technique for equating test scores is implemented, supporting the Equivalent Groups (EG), Single Group (SG), Counterbalanced (CB), Non-Equivalent groups with Anchor Test Chain Equating (NEAT CE), Non-Equivalent groups with Anchor Test Post-Stratification Equating (NEAT PSE) and Non-Equivalent groups with Covariates (NEC) designs. Support for three types of kernels is provided: gaussian, logisitic and uniform. Standard errors of equating and standard errors of the difference between two equating functions are provided for all designs and kernels. Also included are functions aiding the search for a proper log-linear pre-smoothing model and the ability to use Item Response Theory Observed Score Equating (IRT-OSE) in the Kernel Equating framework.
}
\details{
\tabular{ll}{
Package: \tab kequate\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2012-08-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>=2.11.0), graphics, methods, stats\cr
}
}

\author{
Bjorn Andersson, Kenny Branberg and Marie Wiberg

Maintainer: Bjorn Andersson \email{bjorn.andersson@statistik.uu.se}

}
\references{
von Davier, A.A., Holland, P.W., Thayer, D.T. (2004). \emph{The Kernel Method of Test Equating.} Springer-Verlag New York.
}

\keyword{package}