\name{simeq}
\docType{data}
\alias{simeq}
\title{Simulated Test Data}
\description{Contains bivariate test score data and IRT data to be used in equating.}
\usage{data(simeq)}
\format{A list containing two data frames (bivar1 and bivar2) with 1000 observations each and two 5x20 matrices (irtNEATx and irtNEATy).}
\source{The data was generated in R.}
\keyword{datasets}